feat: add more config presets

+++ packages/basic/index.js
@@ -22,7 +22,7 @@ module.exports = {
       files: ['*.json', '*.json5'],
       parser: 'jsonc-eslint-parser',
       rules: {
-        'quotes': ['error', 'double'],
+        'quotes': ['error', 'single'],
         'quote-props': ['error', 'always'],
         'comma-dangle': ['error', 'never'],
       },
